Bug 1893119: Make JS clients of nsIDragService::GetCurrentSession use a (implicit) window r=gstoll,rkraesig,edgar!
GetCurrentSession now needs to know the widget to return the drag session for.
It now uses an explicitly provided window or the current JS context (that must
be a window).

When later patches in the series update nsIWidget::GetDragSession, these
callers will use a drag session that is specific to the widget they supply
here.

This also adds GetDragSession to DOMWindowUtils, and fixes some small bugs
in EventUtils.

Differential Revision: https://phabricator.services.mozilla.com/D211064