Bug 1893119: Remove nsContentUtils::GetDragSession in favor of nsIWidget/BrowsingContext methods r=gstoll,rkraesig,edgar
Updates each client of the nsContentUtils method to get the right drag session -- the one for the widget that is currently the target of the drag session.
The change to nsDOMWindowUtils::DispatchDOMEventViaPresShellForTesting() supports the change to WidgetDragEvent::InitDropEffectForTests() and enabled a
large number of test fixes in the next patch.

Differential Revision: https://phabricator.services.mozilla.com/D211067