Bug 1893119: Make nsBaseDragService weakly remember PBrowsers instead of PContents r=nika!,#win-reviewers!,edgar
nsBaseDragService keeps track of the processes that might need to cancel a drag with EndDragSession and this is technically correct since the drag session is a singleton in the child process, but this series of patches changes those sessions to be per-widget.  This allows content processes to distinguish which of its browsers is engaged in a drag.

Differential Revision: https://phabricator.services.mozilla.com/D211062