Bug 1893119: Move MaybeAddBrowser, RemoveAllBrowsers and MustUpdateDataTransfer from nsIDragService to nsIDragSession and nsIWidget r=gstoll,rkraesig
MustUpdateDataTransfer is a simple function that makes more sense on nsIDragSession.
MaybeAddBrowser records BrowserParents for processes with existing drag sessions.  There may be such processes before the parent process even has a drag session, so they cannot be tied to drag sessions.  We tie them to the widget, until the widget creates a drag session in the process, at which point we move them to the drag session.  We also rename them in the widget case, for clarity, since they are no longer on a drag-specific object.

Differential Revision: https://phabricator.services.mozilla.com/D211081