Bug 1893119: Fix coordinates use in dispatchDOMEventViaPresShellForTesting r=masayuki!,mconley!,edgar!,#places-reviewers!
Since we now add the widget to the event in
dispatchDOMEventViaPresShellForTesting, WidgetGUIEvents that are sent in
mochitests via that method need to transform their screen coordinates by
nsIWidget::WidgetToScreenOffset, to mirror the transformation by
BrowserParent::TransformParentToChild that they don't get because they
skip the parent process.

This exposes a bunch of things that were done to work around this bug.  They
are cleaned up here.

Differential Revision: https://phabricator.services.mozilla.com/D211068