Bug 1893119: Move EndDragSession from nsIDragService to nsIDragSession r=gstoll,rkraesig,edgar
This also adds the widget that hosts the drag session, or a window on
that widget, as a parameter to EndDragSession.  When the service/session
separation happens in a later patch, this will be used to remove the
session from the widget.

Note that EndDragSession now *needs* to be called on for both
the source and target widgets of the drag, if the drag originated
and ended in Gecko.  We have always done this -- it just previously resulted
in double EndDragSession calls (which are harmless).

Differential Revision: https://phabricator.services.mozilla.com/D211083