Bug 1871222: Intercept drop events for content analysis r=smaug!,gstoll!,rkraesig!
Sends drop events to content analysis for approval before sending them to the
DOM.  For in-process drops, we intercept them in
PresShell::EventHandler::HandleEventUsingCoordinates.  For OOP drops, this
first happens in EventStateManager::DispatchCrossProcessEvent, where we detect
the potential need for content analysis (and present a modal dialog on the tab).
The drag session then ends as usual in the parent process but remains open on the
related widget in the process where it would be sent to the DOM while the drop
is being analyzed.  The child process will send either CancelAnalyzeDropEvent,
if CA isn't needed, or AnalyzeDropEvent, if it is.  On AnalyzeDropEvent, CA
will respond by sending ContentAnalysisDropResult back to the child with the
verdict, which will result in either a drop event (on approval) or dragLeave
(otherwise).  Calls to EndDragSession are delayed during this.

Differential Revision: https://phabricator.services.mozilla.com/D211085