Bug 1858871: Part 2 - Inform user of system UX issues initiated by geolocation request r=gstoll!,emilio!
This patch installs the framework for various platforms to inform the user of
one of two things: 1) a coming system dialog asking for system geolocation
permission that will be presented after the Firefox doorhanger asking for
permission for the page, and 2) that we will open a system preferences window,
where the user can enable geolocation for Firefox because it is currently
not enabled.  The code that handles this has been remoted to the parent
process since implementations will not be able to operate in the content
process sandbox.

Here, it stubs the behavior so this does nothing on every platform.
In this patch series, the behavior will be implemented for Windows.

Note: The code will run the geolocation for the page if the user granted it
in Firefox, regardless of whether the user granted or canceled the system
permission.  This respects the user's instruction and provides a
work-around in the event of a bug, although it would usually either fail
to get a location or it will get a very poor one (e.g. via IP lookup).