Bug 1858871: Part 3 - Implement system geolocation permission UX on Windows r=win-reviewers!
This implements PresentSystemSettings, LocationIsPermittedHint and
SystemWillPromptForPermissionHint for Windows.  The Windows APIs are not always
available -- most are currently only available in Windows 11 Canary builds
(slated for September release).  In the event that APIs are not available,
this should do nothing.  At present, this is detailed here:

https://learn.microsoft.com/en-us/windows/win32/nativewifi/wi-fi-access-location-changes

There are two issues that this is intended to handle:

1) The system will display a one-time (or so) dialog to the user when
Firefox requests geolocation but doesn't have permission.  For that
case, we inform the user that they will be asked to grant location permission
again.  This system dialog is only presented in versions of Windows that
support all of the relevant APIs.

2) We open system settings to the right page and post a cancelable modal
dialog on the tab if the user grants geolocation to the page but
geolocation permission isn't currently granted in the OS.  This
case will not happen if case #1 did.  Unfortunately, we can't get
information about the permission status without a location request on old
versions of Windows, so this also does nothing unless the recent APIs are
supported (in this case, AppCapability::CheckAccess).

This work is necessitated not only by the new (occasional) system dialog but
also by Microsoft's plans to block wifi scanning if geolocation isn't
available.  We have used wifi scanning as part of a fallback when system
geolocation isn't available -- that approach is no longer viable here.  A user
would confusingly get repeated errors or very poor results (e.g. IP lookup
results) without information as to why, if that happened.  The fallback
remains useful on other platforms.